+++ vllm_rbln/distributed/kv_transfer/kv_connector/v1/rbln_nixl/__init__.py
@@ -14,17 +14,21 @@
 
 """RBLN NIXL KV-cache transfer connector (mirrors vLLM's v1/nixl/ layout)."""
 
+from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.base_scheduler import (
+    RblnNixlSchedulerBase,
+)
+from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.base_worker import (
+    RblnNixlWorkerBase,
+)
 from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.connector import (
     RblnNixlConnectorBase,
     RblnNixlPullConnector,
 )
-from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.scheduler import (
+from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.pull_scheduler import (
     RblnNixlPullConnectorScheduler,
-    RblnNixlSchedulerBase,
 )
-from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.worker import (
+from vllm_rbln.distributed.kv_transfer.kv_connector.v1.rbln_nixl.pull_worker import (
     RblnNixlPullConnectorWorker,
-    RblnNixlWorkerBase,
 )
 
 __all__ = [
